query-engine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.0
python-multipart==0.0.6
PyYAML==6.0.1
orjson==3.9.10
zstandard==0.22.0